small rearrangements
